fix directory architecture
