+++ src/domain/index.ts
@@ -1,0 +1,2 @@
+export * from './knowledge';
+export * from './model';
