+++ src/domain/knowledge/attributes/index.ts
@@ -1,0 +1,3 @@
+export * from './key';
+export * from './type';
+export * from './mapping';
