+++ src/domain/knowledge/index.ts
@@ -1,0 +1,2 @@
+export * from './attributes';
+export * from './compass';
