+++ src/domain/model/abstract.ts
