+++ src/domain/model/edges.ts
@@ -1,8 +1,8 @@
-import { EdgeAttributes, IAttributes, IEdge, EdgeTargetTuple } from '../types';
 import { DotObject } from './abstract';
-import { attribute } from '../attribute';
+import { EdgeAttributeKey } from '../knowledge';
 import { Attributes } from './attributes-base';
-import { isNodeRefLike } from './utils';
+import { EdgeAttributes, EdgeTargetTuple, IAttributes, IEdge } from './types';
+import { isNodeRefLike } from './nodes';
 
 /**
  * @category Primary
@@ -10,13 +10,13 @@ import { isNodeRefLike } from './utils';
 export class Edge extends DotObject implements IEdge {
   /** Comments to include when outputting with toDot. */
   public comment?: string;
-  public readonly attributes: IAttributes<attribute.Edge>;
+  public readonly attributes: IAttributes<EdgeAttributeKey>;
 
   constructor(public readonly targets: EdgeTargetTuple, attributes?: EdgeAttributes) {
     super();
     if (targets.length < 2 && (isNodeRefLike(targets[0]) && isNodeRefLike(targets[1])) === false) {
       throw Error('The element of Edge target is missing or not satisfied as Edge target.');
     }
-    this.attributes = new Attributes<attribute.Edge>(attributes);
+    this.attributes = new Attributes<EdgeAttributeKey>(attributes);
   }
 }
