+++ src/index.ts
@@ -1,10 +1,2 @@
-export * from './attribute';
-export * from './types';
-export * from './model/abstract';
-export * from './model/attributes-base';
-export * from './model/clusters';
-export * from './model/root-clusters';
-export * from './model/edges';
-export * from './model/nodes';
+export * from './domain';
 export * from './usecase';
-export * from './render/to-dot';
