+++ src/usecase/render/__tests__/attributes.test.ts
